requirements.txt
streamlit
pandas
beautifulsoup4
requests
google-generativeai
lxml